+++ include/bitsery/traits/core/std_defaults.h
@@ -87,7 +87,7 @@ namespace bitsery {
             static void increaseBufferSize(T& container) {
                 //since we're writing to buffer use different resize strategy than default implementation
                 //when small size grow faster, to avoid thouse 2/4/8/16... byte allocations
-                auto newSize = static_cast<size_t>(container.size() * 1.5 + 128);
+                auto newSize = static_cast<size_t>(static_cast<double>(container.size()) * 1.5) + 128;
                 //make data cache friendly
                 newSize -= newSize % 64;//64 is cache line size
                 container.resize((std::max)(newSize, container.capacity()));
